AX: Collapse AccessibilityLabel into AccessibilityRenderObject https://bugs.webkit.org/show_bug.cgi?id=297634 rdar://158730216
Reviewed by Tyler Wilcock.

Collapses all of the logic from AccessibilityLabel into AccessibilityRender- and
AccessibilityNodeObject as applicable.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/accessibility/AXObjectCache.cpp:
(WebCore::AXObjectCache::createObjectFromRenderer):
(WebCore::AXObjectCache::createFromNode):
(WebCore::AXObjectCache::handleChildrenChanged):
(WebCore::AXObjectCache::updateIsolatedTree):
(WebCore::shouldCreateAccessibilityLabel): Deleted.
* Source/WebCore/accessibility/AccessibilityLabel.cpp: Removed.
* Source/WebCore/accessibility/AccessibilityLabel.h: Removed.
* Source/WebCore/accessibility/AccessibilityNodeObject.cpp:
(WebCore::AccessibilityNodeObject::clearChildren):
(WebCore::AccessibilityNodeObject::textAsLabelFor const):
(WebCore::AccessibilityNodeObject::textForLabelElements const):
(WebCore::childrenContainOnlyStaticText):
(WebCore::AccessibilityNodeObject::isLabelContainingOnlyStaticText const):
(WebCore::AccessibilityNodeObject::isNativeLabel const):
* Source/WebCore/accessibility/AccessibilityNodeObject.h:
* Source/WebCore/accessibility/AccessibilityObject.h:
(WebCore::AccessibilityObject::isAccessibilityTableColumnInstance const):
(WebCore::AccessibilityObject::isNativeLabel const):
(WebCore::AccessibilityObject::isLabelByRelation const):
(WebCore::AccessibilityObject::isLabel const):
(WebCore::AccessibilityObject::isAccessibilityLabelInstance const): Deleted.
* Source/WebCore/accessibility/AccessibilityRenderObject.cpp:
(WebCore::AccessibilityRenderObject::stringValue const):
(WebCore::AccessibilityRenderObject::addChildren):
* Source/WebCore/accessibility/isolatedtree/AXIsolatedObject.h:
* Source/WebCore/accessibility/isolatedtree/AXIsolatedTree.cpp:
* Source/WebCore/accessibility/mac/AccessibilityObjectMac.mm:
* Source/WebCore/accessibility/mac/WebAccessibilityObjectWrapperMac.mm:

Canonical link: https://commits.webkit.org/299008@main
