+++ Source/WebCore/accessibility/isolatedtree/AXIsolatedObject.h
@@ -30,6 +30,7 @@
 #include <WebCore/AXCoreObject.h>
 #include <WebCore/AXIsolatedTree.h>
 #include <WebCore/AXObjectCache.h>
+#include <WebCore/AXTreeStoreInlines.h>
 #include <WebCore/IntPoint.h>
 #include <WebCore/LayoutRect.h>
 #include <WebCore/NodeName.h>
