+++ Source/WebCore/accessibility/mac/WebAccessibilityObjectWrapperMac.mm
@@ -42,7 +42,6 @@
 #import "AXTextMarker.h"
 #import "AXTreeStore.h"
 #import "AXTreeStoreInlines.h"
-#import "AccessibilityLabel.h"
 #import "AccessibilityProgressIndicator.h"
 #import "AccessibilityRenderObject.h"
 #import "AccessibilityScrollView.h"
